requirements.txt
torch>=2.4.0
torchvision>=0.19.0
transformers>=4.44.0
accelerate>=0.33.0
datasets>=2.20.0
huggingface_hub>=0.24.0
safetensors>=0.4.3
wandb>=0.17.0
triton>=3.0.0
tqdm>=4.66.0
numpy>=1.26.0
scipy>=1.13.0
matplotlib>=3.9.0
sentencepiece>=0.2.0
tiktoken>=0.7.0
einops>=0.8.0
pyyaml>=6.0.1
